move tube handler functions to tube module so components can stay separate
